+++ src/main/java/com/programmers/bucketback/domains/member/application/MemberService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 
 import com.programmers.bucketback.domains.common.MemberUtils;
 import com.programmers.bucketback.domains.member.application.dto.response.LoginMemberServiceResponse;
-import com.programmers.bucketback.domains.member.domain.LoginInfo;
 import com.programmers.bucketback.domains.member.domain.Member;
+import com.programmers.bucketback.domains.member.domain.vo.LoginInfo;
 import com.programmers.bucketback.global.error.exception.BusinessException;
 import com.programmers.bucketback.global.error.exception.ErrorCode;
