--- src/main/java/com/programmers/bucketback/domains/member/domain/MemberStatus.java
@@ -1,6 +1,0 @@
-package com.programmers.bucketback.domains.member.domain;
-
-public enum MemberStatus {
-	ACTIVE,
-	DELETED
-}
