--- src/main/java/com/programmers/bucketback/domains/member/domain/Role.java
@@ -1,6 +1,0 @@
-package com.programmers.bucketback.domains.member.domain;
-
-public enum Role {
-	USER,
-	ADMIN
-}
