+++ src/main/java/com/programmers/bucketback/domains/member/domain/vo/MemberStatus.java
@@ -1,0 +1,6 @@
+package com.programmers.bucketback.domains.member.domain.vo;
+
+public enum MemberStatus {
+	ACTIVE,
+	DELETED
+}
