+++ src/main/java/com/programmers/bucketback/domains/member/domain/vo/Role.java
@@ -1,0 +1,6 @@
+package com.programmers.bucketback.domains.member.domain.vo;
+
+public enum Role {
+	USER,
+	ADMIN
+}
